updated icons & reusable components
