+++ lib/core/constants.dart
@@ -1,0 +1,19 @@
+import 'package:portfolio/models/portfolio_models.dart';
+
+final socialLinks = [
+  const SocialLink(
+    name: 'LinkedIn',
+    icon: 'assets/icons/linkedin_icon.svg',
+    url: 'https://www.linkedin.com/in/amarjit-mallick/',
+  ),
+  const SocialLink(
+    name: 'GitHub',
+    icon: 'assets/icons/github_icon.svg',
+    url: 'https://github.com/amarjitmallick',
+  ),
+  const SocialLink(
+    name: 'Twitter',
+    icon: 'assets/icons/x_icon.svg',
+    url: 'https://x.com/amarjitmallick_',
+  ),
+];
