+++ lib/screens/about_screen.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:portfolio/data/portfolio_data.dart';
 import 'package:portfolio/models/portfolio_models.dart';
 import 'package:portfolio/widgets/animated_section.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:portfolio/widgets/profile_hero.dart';
+import 'package:portfolio/widgets/section_header.dart';
+import 'package:portfolio/widgets/skills_section.dart';
+import 'package:portfolio/widgets/timeline_card.dart';
 
 class AboutScreen extends StatelessWidget {
   const AboutScreen({super.key});
@@ -19,6 +20,24 @@ class AboutScreen extends StatelessWidget {
     final isTablet = screenSize.width >= 768 && screenSize.width < 1024;
     final isMobile = screenSize.width < 768;
 
+    final socialLinks = [
+      const SocialLink(
+        name: 'LinkedIn',
+        icon: 'assets/icons/linkedin_icon.svg',
+        url: 'https://www.linkedin.com/in/amarjit-mallick/',
+      ),
+      const SocialLink(
+        name: 'GitHub',
+        icon: 'assets/icons/github_icon.svg',
+        url: 'https://github.com/amarjitmallick',
+      ),
+      const SocialLink(
+        name: 'Twitter',
+        icon: 'assets/icons/x_icon.svg',
+        url: 'https://x.com/amarjitmallick_',
+      ),
+    ];
+
     return Scaffold(
       body: Center(
         child: Column(
@@ -31,7 +50,10 @@ class AboutScreen extends StatelessWidget {
                     flex: 2,
                     child: AnimatedSection(
                       delay: const Duration(milliseconds: 200),
-                      child: _buildHeroImage(context, personalInfo),
+                      child: ProfileHero(
+                        personalInfo: personalInfo,
+                        socialLinks: socialLinks,
+                      ),
                     ),
                   ),
                   Expanded(
@@ -55,7 +77,22 @@ class AboutScreen extends StatelessWidget {
                             children: [
                               AnimatedSection(
                                 delay: const Duration(milliseconds: 200),
-                                child: _buildHeader(context),
+                                child: Column(
+                                  crossAxisAlignment: CrossAxisAlignment.start,
+                                  children: [
+                                    const SectionHeader(
+                                      title: "ABOUT ME",
+                                      subtitle: "& SKILLS",
+                                    ),
+                                    const SizedBox(height: 8),
+                                    Text(
+                                      isMobile
+                                          ? 'Get to know me, my skills, and professional journey'
+                                          : 'Get to know me better and explore my technical expertise',
+                                      style: const TextStyle(fontSize: 18),
+                                    ),
+                                  ],
+                                ),
                               ),
                               SizedBox(
                                 height: isDesktop
@@ -82,12 +119,7 @@ class AboutScreen extends StatelessWidget {
                               ),
                               AnimatedSection(
                                 delay: const Duration(milliseconds: 600),
-                                child: _buildSkillsSection(
-                                  context,
-                                  skills,
-                                  isDesktop,
-                                  isTablet,
-                                ),
+                                child: SkillsSection(skills: skills),
                               ),
                               SizedBox(
                                 height: isDesktop
@@ -106,11 +138,51 @@ class AboutScreen extends StatelessWidget {
                                 ),
                                 AnimatedSection(
                                   delay: const Duration(milliseconds: 800),
-                                  child: _buildCareerSection(
-                                    context,
-                                    careerItems,
-                                    isDesktop,
-                                    isTablet,
+                                  child: Column(
+                                    crossAxisAlignment:
+                                        CrossAxisAlignment.start,
+                                    children: [
+                                      Row(
+                                        children: [
+                                          Icon(
+                                            Icons.timeline,
+                                            color: Theme.of(
+                                              context,
+                                            ).colorScheme.primary,
+                                          ),
+                                          const SizedBox(width: 8),
+                                          Text(
+                                            'Career Journey',
+                                            style: Theme.of(context)
+                                                .textTheme
+                                                .headlineMedium
+                                                ?.copyWith(
+                                                  fontWeight: FontWeight.bold,
+                                                ),
+                                          ),
+                                        ],
+                                      ),
+                                      const SizedBox(height: 24),
+                                      ...careerItems.asMap().entries.map((
+                                        entry,
+                                      ) {
+                                        final index = entry.key;
+                                        final item = entry.value;
+                                        return AnimatedSection(
+                                          delay: Duration(
+                                            milliseconds: 200 * index,
+                                          ),
+                                          child: Padding(
+                                            padding: EdgeInsets.only(
+                                              bottom: isTablet ? 24 : 20,
+                                            ),
+                                            child: TimelineCard(
+                                              careerItem: item,
+                                            ),
+                                          ),
+                                        );
+                                      }),
+                                    ],
                                   ),
                                 ),
                               ],
@@ -129,190 +201,9 @@ class AboutScreen extends StatelessWidget {
     );
   }
 
-  Widget _buildHeroImage(BuildContext context, PersonalInfo personalInfo) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.white.withValues(alpha: 0.1),
-        borderRadius: BorderRadius.circular(32),
-      ),
-      margin: EdgeInsets.symmetric(vertical: 50, horizontal: 80),
-      padding: EdgeInsets.all(40),
-      constraints: const BoxConstraints(maxWidth: 300),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          AspectRatio(
-            aspectRatio: 1,
-            child: Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(24),
-                boxShadow: [
-                  BoxShadow(
-                    color: Theme.of(
-                      context,
-                    ).colorScheme.shadow.withValues(alpha: 0.1),
-                    blurRadius: 20,
-                    offset: const Offset(0, 10),
-                  ),
-                ],
-              ),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(24),
-                child: Image.asset(
-                  personalInfo.profileImage,
-                  fit: BoxFit.cover,
-                  errorBuilder: (context, error, stackTrace) {
-                    return Container(
-                      decoration: BoxDecoration(
-                        color: Theme.of(context).colorScheme.primaryContainer,
-                        borderRadius: BorderRadius.circular(20),
-                      ),
-                      child: Icon(
-                        Icons.person,
-                        size: 100,
-                        color: Theme.of(context).colorScheme.onPrimaryContainer,
-                      ),
-                    );
-                  },
-                ),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 20.0),
-            child: Text(
-              "Amarjit Mallick",
-              style: TextStyle(
-                fontSize: 42,
-                fontWeight: FontWeight.w800,
-                letterSpacing: 2,
-                color: Colors.white,
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 20.0),
-            child: Text(
-              "A Software Engineer who has developed countless innovative solutions",
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 18,
-              ),
-            ),
-          ),
-          _buildSocialLinks(context),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSocialLinks(
-    BuildContext context,
-  ) {
-    final socialLinks = [
-      {
-        'name': 'LinkedIn',
-        'icon': 'assets/icons/linkedin_icon.svg',
-        'url': 'https://www.linkedin.com/in/amarjit-mallick/',
-      },
-      {
-        'name': 'GitHub',
-        'icon': 'assets/icons/github_icon.svg',
-        'url': 'https://github.com/amarjitmallick',
-      },
-      {
-        'name': 'Twitter',
-        'icon': 'assets/icons/x_icon.svg',
-        'url': 'https://x.com/amarjitmallick_',
-      },
-    ];
-
-    return Container(
-      constraints: const BoxConstraints(maxWidth: 600),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            spacing: 16,
-            children: socialLinks.map((link) {
-              return IconButton(
-                style: IconButton.styleFrom(
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                ),
-                // tooltip: link['name'],
-                onPressed: () {
-                  launchUrl(
-                    Uri.parse(link['url'] as String),
-                    mode: LaunchMode.externalApplication,
-                  );
-                },
-                icon: SvgPicture.asset(
-                  link['icon'] as String,
-                  width: 36,
-                  height: 36,
-                  fit: BoxFit.contain,
-                  colorFilter: ColorFilter.mode(Colors.white, BlendMode.srcIn),
-                ),
-              );
-            }).toList(),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildHeader(BuildContext context) {
-    final screenSize = MediaQuery.of(context).size;
-    final isMobile = screenSize.width < 768;
-
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        AnimatedSection(
-          delay: const Duration(milliseconds: 300),
-          child: RichText(
-            text: TextSpan(
-              text: "ABOUT ME\n",
-              style: GoogleFonts.urbanist().copyWith(
-                fontSize: 120,
-                fontWeight: FontWeight.w600,
-                letterSpacing: 2,
-                color: Colors.white,
-              ),
-              children: [
-                TextSpan(
-                  text: "& SKILLS",
-                  style: TextStyle(
-                    fontSize: 120,
-                    fontWeight: FontWeight.bold,
-                    letterSpacing: 2,
-                    color: Colors.white.withValues(
-                      alpha: 0.35,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ),
-        const SizedBox(height: 8),
-        Text(
-          isMobile
-              ? 'Get to know me, my skills, and professional journey'
-              : 'Get to know me better and explore my technical expertise',
-          style: TextStyle(fontSize: 18),
-        ),
-      ],
-    );
-  }
-
   Widget _buildAboutContent(
     BuildContext context,
-    personalInfo,
+    PersonalInfo personalInfo,
     bool isDesktop,
     bool isTablet,
   ) {
@@ -336,7 +227,7 @@ class AboutScreen extends StatelessWidget {
     );
   }
 
-  Widget _buildAboutText(BuildContext context, personalInfo) {
+  Widget _buildAboutText(BuildContext context, PersonalInfo personalInfo) {
     return Container(
       decoration: BoxDecoration(
         color: Theme.of(context).colorScheme.surfaceContainerHighest,
@@ -378,7 +269,7 @@ class AboutScreen extends StatelessWidget {
     );
   }
 
-  Widget _buildPersonalInfo(BuildContext context, personalInfo) {
+  Widget _buildPersonalInfo(BuildContext context, PersonalInfo personalInfo) {
     return Container(
       padding: const EdgeInsets.all(24),
       decoration: BoxDecoration(
@@ -469,300 +360,6 @@ class AboutScreen extends StatelessWidget {
                   style: Theme.of(context).textTheme.bodyMedium?.copyWith(
                     fontWeight: FontWeight.w500,
                   ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSkillsSection(
-    BuildContext context,
-    List skills,
-    bool isDesktop,
-    bool isTablet,
-  ) {
-    final groupedSkills = <String, List>{};
-    for (final skill in skills) {
-      groupedSkills.putIfAbsent(skill.category, () => []).add(skill);
-    }
-
-    return Container(
-      decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.surfaceContainerHighest,
-        borderRadius: BorderRadius.circular(16),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(
-            'Technical Skills',
-            style: Theme.of(context).textTheme.titleLarge?.copyWith(
-              fontWeight: FontWeight.w600,
-              fontSize: 32,
-            ),
-          ),
-          const SizedBox(height: 24),
-          Table(
-            border: TableBorder.all(
-              color: Colors.white.withValues(alpha: 0.2),
-              borderRadius: BorderRadius.circular(8),
-            ),
-            children: [
-              ...groupedSkills.entries.map((entry) {
-                return TableRow(
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 12,
-                        vertical: 12,
-                      ),
-                      child: Text(
-                        entry.key,
-                        style: Theme.of(context).textTheme.titleMedium
-                            ?.copyWith(fontWeight: FontWeight.w600),
-                      ),
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 12,
-                        vertical: 12,
-                      ),
-                      child: Row(
-                        spacing: 10,
-                        children: entry.value
-                            .map(
-                              (skill) => Text(
-                                skill.name,
-                                style: Theme.of(context).textTheme.titleMedium
-                                    ?.copyWith(fontWeight: FontWeight.w600),
-                              ),
-                            )
-                            .toList(),
-                      ),
-                    ),
-                  ],
-                );
-              }),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSkillCategory(
-    BuildContext context,
-    String category,
-    List categorySkills,
-    bool isDesktop,
-    bool isTablet,
-  ) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          category,
-          style: Theme.of(context).textTheme.titleLarge?.copyWith(
-            fontWeight: FontWeight.w600,
-            color: Theme.of(context).colorScheme.primary,
-          ),
-        ),
-        SizedBox(height: isDesktop ? 20 : 16),
-        SizedBox(
-          height: 50,
-          child: ListView.separated(
-            scrollDirection: Axis.horizontal,
-            physics: const NeverScrollableScrollPhysics(),
-            itemCount: categorySkills.length,
-            itemBuilder: (context, index) {
-              final skill = categorySkills[index];
-              return _buildSkillItem(context, skill);
-            },
-            separatorBuilder: (BuildContext context, int index) {
-              return SizedBox(
-                width: 10,
-              );
-            },
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildSkillItem(BuildContext context, skill) {
-    return Chip(
-      label: Text(
-        skill.name,
-        style: Theme.of(context).textTheme.titleMedium?.copyWith(
-          fontWeight: FontWeight.w600,
-        ),
-      ),
-    );
-  }
-
-  Widget _buildCareerSection(
-    BuildContext context,
-    List careerItems,
-    bool isDesktop,
-    bool isTablet,
-  ) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Icon(
-              Icons.timeline,
-              color: Theme.of(context).colorScheme.primary,
-            ),
-            const SizedBox(width: 8),
-            Text(
-              'Career Journey',
-              style: Theme.of(context).textTheme.headlineMedium?.copyWith(
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 24),
-        ...careerItems.asMap().entries.map((entry) {
-          final index = entry.key;
-          final item = entry.value;
-          return AnimatedSection(
-            delay: Duration(milliseconds: 200 * index),
-            child: Padding(
-              padding: EdgeInsets.only(bottom: isTablet ? 24 : 20),
-              child: _buildCareerItem(context, item),
-            ),
-          );
-        }),
-      ],
-    );
-  }
-
-  Widget _buildCareerItem(BuildContext context, item) {
-    return Container(
-      padding: const EdgeInsets.all(20),
-      decoration: BoxDecoration(
-        color: Theme.of(context).colorScheme.surfaceContainerHighest,
-        border: Border.all(
-          color: Theme.of(context).colorScheme.outline.withValues(alpha: 0.2),
-        ),
-        borderRadius: BorderRadius.circular(16),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            children: [
-              Container(
-                width: 8,
-                height: 8,
-                decoration: BoxDecoration(
-                  color: Theme.of(context).colorScheme.primary,
-                  shape: BoxShape.circle,
-                ),
-              ),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Text(
-                  item.position,
-                  style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                    fontWeight: FontWeight.w600,
-                  ),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          Padding(
-            padding: const EdgeInsets.only(left: 20),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  item.company,
-                  style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                    color: Theme.of(context).colorScheme.primary,
-                    fontWeight: FontWeight.w500,
-                  ),
-                ),
-                const SizedBox(height: 4),
-                Text(
-                  item.duration,
-                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Theme.of(
-                      context,
-                    ).colorScheme.onSurface.withValues(alpha: 0.7),
-                  ),
-                ),
-                const SizedBox(height: 12),
-                Text(
-                  item.description,
-                  style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    height: 1.5,
-                  ),
-                ),
-                const SizedBox(height: 20),
-                Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Row(
-                      children: [
-                        Icon(
-                          Icons.star,
-                          size: 16,
-                          color: Theme.of(context).colorScheme.primary,
-                        ),
-                        const SizedBox(width: 8),
-                        Text(
-                          'Key Achievements',
-                          style: Theme.of(context).textTheme.titleSmall
-                              ?.copyWith(
-                                fontWeight: FontWeight.w600,
-                                color: Theme.of(context).colorScheme.onSurface,
-                              ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 12),
-                    ...item.achievements.map<Widget>((achievement) {
-                      return Padding(
-                        padding: const EdgeInsets.only(bottom: 8),
-                        child: Row(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Container(
-                              margin: const EdgeInsets.only(top: 6),
-                              width: 4,
-                              height: 4,
-                              decoration: BoxDecoration(
-                                color: Theme.of(context).colorScheme.primary,
-                                shape: BoxShape.circle,
-                              ),
-                            ),
-                            const SizedBox(width: 12),
-                            Expanded(
-                              child: Text(
-                                achievement,
-                                style: Theme.of(context).textTheme.bodyMedium
-                                    ?.copyWith(
-                                      height: 1.5,
-                                      color: Theme.of(context)
-                                          .colorScheme
-                                          .onSurface
-                                          .withValues(alpha: 0.8),
-                                    ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      );
-                    }),
-                  ],
                 ),
               ],
             ),
