+++ lib/screens/career_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
+import 'package:portfolio/core/constants.dart';
 import 'package:portfolio/data/portfolio_data.dart';
 import 'package:portfolio/widgets/animated_section.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:portfolio/widgets/profile_hero.dart';
 
 import '../models/portfolio_models.dart';
 
@@ -29,7 +30,10 @@ class CareerScreen extends StatelessWidget {
                     flex: 2,
                     child: AnimatedSection(
                       delay: const Duration(milliseconds: 200),
-                      child: _buildHeroImage(context, personalInfo),
+                      child: ProfileHero(
+                        personalInfo: personalInfo,
+                        socialLinks: socialLinks,
+                      ),
                     ),
                   ),
                   Expanded(
@@ -44,7 +48,9 @@ class CareerScreen extends StatelessWidget {
                       ),
                       child: Center(
                         child: ConstrainedBox(
-                          constraints: BoxConstraints(maxWidth: isDesktop ? 1400 : 900),
+                          constraints: BoxConstraints(
+                            maxWidth: isDesktop ? 1400 : 900,
+                          ),
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
                             children: [
@@ -59,7 +65,12 @@ class CareerScreen extends StatelessWidget {
                                     ? 32
                                     : 24,
                               ),
-                              _buildTimeline(context, careerItems, isDesktop, isTablet),
+                              _buildTimeline(
+                                context,
+                                careerItems,
+                                isDesktop,
+                                isTablet,
+                              ),
                             ],
                           ),
                         ),
@@ -71,135 +82,6 @@ class CareerScreen extends StatelessWidget {
             ),
           ],
         ),
-      ),
-    );
-  }
-
-  Widget _buildHeroImage(BuildContext context, PersonalInfo personalInfo) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.white.withValues(alpha: 0.1),
-        borderRadius: BorderRadius.circular(32),
-      ),
-      margin: EdgeInsets.symmetric(vertical: 50, horizontal: 80),
-      padding: EdgeInsets.all(40),
-      constraints: const BoxConstraints(maxWidth: 300),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          AspectRatio(
-            aspectRatio: 1,
-            child: Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(24),
-                boxShadow: [
-                  BoxShadow(
-                    color: Theme.of(
-                      context,
-                    ).colorScheme.shadow.withValues(alpha: 0.1),
-                    blurRadius: 20,
-                    offset: const Offset(0, 10),
-                  ),
-                ],
-              ),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(24),
-                child: Image.asset(
-                  personalInfo.profileImage,
-                  fit: BoxFit.cover,
-                  errorBuilder: (context, error, stackTrace) {
-                    return Container(
-                      decoration: BoxDecoration(
-                        color: Theme.of(context).colorScheme.primaryContainer,
-                        borderRadius: BorderRadius.circular(20),
-                      ),
-                      child: Icon(
-                        Icons.person,
-                        size: 100,
-                        color: Theme.of(context).colorScheme.onPrimaryContainer,
-                      ),
-                    );
-                  },
-                ),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 20.0),
-            child: Text(
-              "Amarjit Mallick",
-              style: TextStyle(
-                fontSize: 42,
-                fontWeight: FontWeight.w800,
-                letterSpacing: 2,
-                color: Colors.white,
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 20.0),
-            child: Text(
-              "A Software Engineer who has developed countless innovative solutions",
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 18,
-              ),
-            ),
-          ),
-          _buildSocialLinks(context),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSocialLinks(
-    BuildContext context,
-  ) {
-    final socialLinks = [
-      {
-        'name': 'LinkedIn',
-        'icon': 'assets/icons/linkedin.png',
-        'url': 'https://www.linkedin.com/in/amarjit-mallick/',
-      },
-      {
-        'name': 'GitHub',
-        'icon': 'assets/icons/github.png',
-        'url': 'https://github.com/amarjitmallick',
-      },
-      {
-        'name': 'Twitter',
-        'icon': 'assets/icons/twitter.png',
-        'url': 'https://x.com/amarjitmallick_',
-      },
-    ];
-
-    return Container(
-      constraints: const BoxConstraints(maxWidth: 600),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            spacing: 16,
-            children: socialLinks.map((link) {
-              return IconButton(
-                onPressed: () {
-                  launchUrl(
-                    Uri.parse(link['url'] as String),
-                    mode: LaunchMode.externalApplication,
-                  );
-                },
-                icon: Image.asset(
-                  link['icon'] as String,
-                  width: 36,
-                  height: 36,
-                  fit: BoxFit.contain,
-                ),
-              );
-            }).toList(),
-          ),
-        ],
       ),
     );
   }
@@ -234,14 +116,21 @@ class CareerScreen extends StatelessWidget {
         Text(
           'My professional experience and career progression',
           style: Theme.of(context).textTheme.bodyLarge?.copyWith(
-            color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.7),
+            color: Theme.of(
+              context,
+            ).colorScheme.onSurface.withValues(alpha: 0.7),
           ),
         ),
       ],
     );
   }
 
-  Widget _buildTimeline(BuildContext context, List<CareerItem> careerItems, bool isDesktop, bool isTablet) {
+  Widget _buildTimeline(
+    BuildContext context,
+    List<CareerItem> careerItems,
+    bool isDesktop,
+    bool isTablet,
+  ) {
     return ListView.builder(
       shrinkWrap: true,
       physics: const NeverScrollableScrollPhysics(),
@@ -258,7 +147,11 @@ class CareerScreen extends StatelessWidget {
     );
   }
 
-  Widget _buildTimelineItem(BuildContext context, CareerItem careerItem, bool isLast) {
+  Widget _buildTimelineItem(
+    BuildContext context,
+    CareerItem careerItem,
+    bool isLast,
+  ) {
     return IntrinsicHeight(
       child: Row(
         crossAxisAlignment: CrossAxisAlignment.start,
@@ -278,7 +171,9 @@ class CareerScreen extends StatelessWidget {
                   ),
                   boxShadow: [
                     BoxShadow(
-                      color: Theme.of(context).colorScheme.primary.withValues(alpha: 0.3),
+                      color: Theme.of(
+                        context,
+                      ).colorScheme.primary.withValues(alpha: 0.3),
                       blurRadius: 8,
                       spreadRadius: 2,
                     ),
@@ -289,7 +184,9 @@ class CareerScreen extends StatelessWidget {
                 Expanded(
                   child: Container(
                     width: 2,
-                    color: Theme.of(context).colorScheme.outline.withValues(alpha: 0.3),
+                    color: Theme.of(
+                      context,
+                    ).colorScheme.outline.withValues(alpha: 0.3),
                     margin: const EdgeInsets.symmetric(vertical: 8),
                   ),
                 ),
@@ -305,11 +202,15 @@ class CareerScreen extends StatelessWidget {
                 color: Theme.of(context).colorScheme.surface,
                 borderRadius: BorderRadius.circular(16),
                 border: Border.all(
-                  color: Theme.of(context).colorScheme.outline.withValues(alpha: 0.2),
+                  color: Theme.of(
+                    context,
+                  ).colorScheme.outline.withValues(alpha: 0.2),
                 ),
                 boxShadow: [
                   BoxShadow(
-                    color: Theme.of(context).colorScheme.shadow.withValues(alpha: 0.05),
+                    color: Theme.of(
+                      context,
+                    ).colorScheme.shadow.withValues(alpha: 0.05),
                     blurRadius: 10,
                     offset: const Offset(0, 4),
                   ),
@@ -326,34 +227,46 @@ class CareerScreen extends StatelessWidget {
                           children: [
                             Text(
                               careerItem.position,
-                              style: Theme.of(context).textTheme.titleLarge?.copyWith(
-                                fontWeight: FontWeight.bold,
-                                color: Theme.of(context).colorScheme.onSurface,
-                              ),
+                              style: Theme.of(context).textTheme.titleLarge
+                                  ?.copyWith(
+                                    fontWeight: FontWeight.bold,
+                                    color: Theme.of(
+                                      context,
+                                    ).colorScheme.onSurface,
+                                  ),
                             ),
                             const SizedBox(height: 4),
                             Text(
                               careerItem.company,
-                              style: Theme.of(context).textTheme.titleMedium?.copyWith(
-                                color: Theme.of(context).colorScheme.primary,
-                                fontWeight: FontWeight.w600,
-                              ),
+                              style: Theme.of(context).textTheme.titleMedium
+                                  ?.copyWith(
+                                    color: Theme.of(
+                                      context,
+                                    ).colorScheme.primary,
+                                    fontWeight: FontWeight.w600,
+                                  ),
                             ),
                           ],
                         ),
                       ),
                       Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
+                        padding: const EdgeInsets.symmetric(
+                          horizontal: 12,
+                          vertical: 6,
+                        ),
                         decoration: BoxDecoration(
                           color: Theme.of(context).colorScheme.primaryContainer,
                           borderRadius: BorderRadius.circular(12),
                         ),
                         child: Text(
                           careerItem.duration,
-                          style: Theme.of(context).textTheme.labelMedium?.copyWith(
-                            color: Theme.of(context).colorScheme.onPrimaryContainer,
-                            fontWeight: FontWeight.w600,
-                          ),
+                          style: Theme.of(context).textTheme.labelMedium
+                              ?.copyWith(
+                                color: Theme.of(
+                                  context,
+                                ).colorScheme.onPrimaryContainer,
+                                fontWeight: FontWeight.w600,
+                              ),
                         ),
                       ),
                     ],
@@ -364,7 +277,9 @@ class CareerScreen extends StatelessWidget {
                       careerItem.description,
                       style: Theme.of(context).textTheme.bodyLarge?.copyWith(
                         height: 1.6,
-                        color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.8),
+                        color: Theme.of(
+                          context,
+                        ).colorScheme.onSurface.withValues(alpha: 0.8),
                       ),
                     ),
                   ],
@@ -384,10 +299,13 @@ class CareerScreen extends StatelessWidget {
                             const SizedBox(width: 8),
                             Text(
                               'Key Achievements',
-                              style: Theme.of(context).textTheme.titleSmall?.copyWith(
-                                fontWeight: FontWeight.w600,
-                                color: Theme.of(context).colorScheme.onSurface,
-                              ),
+                              style: Theme.of(context).textTheme.titleSmall
+                                  ?.copyWith(
+                                    fontWeight: FontWeight.w600,
+                                    color: Theme.of(
+                                      context,
+                                    ).colorScheme.onSurface,
+                                  ),
                             ),
                           ],
                         ),
@@ -403,7 +321,9 @@ class CareerScreen extends StatelessWidget {
                                   width: 4,
                                   height: 4,
                                   decoration: BoxDecoration(
-                                    color: Theme.of(context).colorScheme.primary,
+                                    color: Theme.of(
+                                      context,
+                                    ).colorScheme.primary,
                                     shape: BoxShape.circle,
                                   ),
                                 ),
@@ -411,10 +331,16 @@ class CareerScreen extends StatelessWidget {
                                 Expanded(
                                   child: Text(
                                     achievement,
-                                    style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                                      height: 1.5,
-                                      color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.8),
-                                    ),
+                                    style: Theme.of(context)
+                                        .textTheme
+                                        .bodyMedium
+                                        ?.copyWith(
+                                          height: 1.5,
+                                          color: Theme.of(context)
+                                              .colorScheme
+                                              .onSurface
+                                              .withValues(alpha: 0.8),
+                                        ),
                                   ),
                                 ),
                               ],
