+++ lib/screens/publications_screen.dart
@@ -1,9 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:intl/intl.dart';
+import 'package:portfolio/core/constants.dart';
 import 'package:portfolio/data/portfolio_data.dart';
 import 'package:portfolio/data/repository/blog_repository.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:portfolio/widgets/profile_hero.dart';
 
 import '../models/portfolio_models.dart';
 import '../widgets/animated_section.dart';
@@ -32,84 +33,108 @@ class PublicationsScreen extends StatelessWidget {
                     flex: 2,
                     child: AnimatedSection(
                       delay: const Duration(milliseconds: 200),
-                      child: _buildHeroImage(context, personalInfo),
+                      child: ProfileHero(
+                        personalInfo: personalInfo,
+                        socialLinks: socialLinks,
+                      ),
                     ),
                   ),
                   Expanded(
                     flex: 3,
                     child: FutureBuilder<List<dynamic>>(
                       future: BlogRepository().fetchBlogs(),
-                      builder: (BuildContext context, AsyncSnapshot<List<dynamic>> snapshot) {
-                        if (snapshot.connectionState == ConnectionState.waiting) {
-                          return Center(
-                            child: CircularProgressIndicator(),
-                          );
-                        } else if (snapshot.hasData) {
-                          if (snapshot.data != null && snapshot.data!.isNotEmpty) {
-                            publications = snapshot.data!.map((blog) {
-                              return Publication(
-                                title: blog["node"]["title"],
-                                description: blog["node"]["brief"],
-                                blogUrl: blog["node"]["url"],
-                                imageUrl: blog["node"]["coverImage"]["url"],
-                                year: blog["node"]["publishedAt"],
+                      builder:
+                          (
+                            BuildContext context,
+                            AsyncSnapshot<List<dynamic>> snapshot,
+                          ) {
+                            if (snapshot.connectionState ==
+                                ConnectionState.waiting) {
+                              return Center(
+                                child: CircularProgressIndicator(),
                               );
-                            }).toList();
-                            return SingleChildScrollView(
-                              padding: EdgeInsets.all(
-                                isDesktop
-                                    ? 40
-                                    : isTablet
-                                    ? 24
-                                    : 16,
-                              ),
-                              child: Center(
-                                child: ConstrainedBox(
-                                  constraints: BoxConstraints(maxWidth: isDesktop ? 1400 : 900),
-                                  child: Column(
-                                    crossAxisAlignment: CrossAxisAlignment.start,
-                                    children: [
-                                      AnimatedSection(
-                                        delay: const Duration(milliseconds: 200),
-                                        child: _buildHeader(context),
-                                      ),
-                                      SizedBox(
-                                        height: isDesktop
-                                            ? 50
-                                            : isTablet
-                                            ? 32
-                                            : 24,
-                                      ),
-                                      ListView.builder(
-                                        shrinkWrap: true,
-                                        physics: const NeverScrollableScrollPhysics(),
-                                        itemCount: publications.length,
-                                        itemBuilder: (context, index) {
-                                          return AnimatedSection(
-                                            delay: Duration(milliseconds: 400 + (index * 200)),
-                                            child: Padding(
-                                              padding: const EdgeInsets.only(bottom: 24),
-                                              child: _buildPublicationCard(context, publications[index]),
-                                            ),
-                                          );
-                                        },
-                                      ),
-                                    ],
+                            } else if (snapshot.hasData) {
+                              if (snapshot.data != null &&
+                                  snapshot.data!.isNotEmpty) {
+                                publications = snapshot.data!.map((blog) {
+                                  return Publication(
+                                    title: blog["node"]["title"],
+                                    description: blog["node"]["brief"],
+                                    blogUrl: blog["node"]["url"],
+                                    imageUrl: blog["node"]["coverImage"]["url"],
+                                    year: blog["node"]["publishedAt"],
+                                  );
+                                }).toList();
+                                return SingleChildScrollView(
+                                  padding: EdgeInsets.all(
+                                    isDesktop
+                                        ? 40
+                                        : isTablet
+                                        ? 24
+                                        : 16,
                                   ),
-                                ),
-                              ),
-                            );
-                          } else {
-                            return Center(
-                              child: Text("No Blogs to show"),
-                            );
-                          }
-                        } else {
-                          return Center(
-                            child: Text("Something went wrong"),
-                          );
-                        }
-                      },
+                                  child: Center(
+                                    child: ConstrainedBox(
+                                      constraints: BoxConstraints(
+                                        maxWidth: isDesktop ? 1400 : 900,
+                                      ),
+                                      child: Column(
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.start,
+                                        children: [
+                                          AnimatedSection(
+                                            delay: const Duration(
+                                              milliseconds: 200,
+                                            ),
+                                            child: _buildHeader(context),
+                                          ),
+                                          SizedBox(
+                                            height: isDesktop
+                                                ? 50
+                                                : isTablet
+                                                ? 32
+                                                : 24,
+                                          ),
+                                          ListView.builder(
+                                            shrinkWrap: true,
+                                            physics:
+                                                const NeverScrollableScrollPhysics(),
+                                            itemCount: publications.length,
+                                            itemBuilder: (context, index) {
+                                              return AnimatedSection(
+                                                delay: Duration(
+                                                  milliseconds:
+                                                      400 + (index * 200),
+                                                ),
+                                                child: Padding(
+                                                  padding:
+                                                      const EdgeInsets.only(
+                                                        bottom: 24,
+                                                      ),
+                                                  child: _buildPublicationCard(
+                                                    context,
+                                                    publications[index],
+                                                  ),
+                                                ),
+                                              );
+                                            },
+                                          ),
+                                        ],
+                                      ),
+                                    ),
+                                  ),
+                                );
+                              } else {
+                                return Center(
+                                  child: Text("No Blogs to show"),
+                                );
+                              }
+                            } else {
+                              return Center(
+                                child: Text("Something went wrong"),
+                              );
+                            }
+                          },
                     ),
                   ),
                 ],
@@ -117,135 +142,6 @@ class PublicationsScreen extends StatelessWidget {
             ),
           ],
         ),
-      ),
-    );
-  }
-
-  Widget _buildHeroImage(BuildContext context, PersonalInfo personalInfo) {
-    return Container(
-      decoration: BoxDecoration(
-        color: Colors.white.withValues(alpha: 0.1),
-        borderRadius: BorderRadius.circular(32),
-      ),
-      margin: EdgeInsets.symmetric(vertical: 50, horizontal: 80),
-      padding: EdgeInsets.all(40),
-      constraints: const BoxConstraints(maxWidth: 300),
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          AspectRatio(
-            aspectRatio: 1,
-            child: Container(
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(24),
-                boxShadow: [
-                  BoxShadow(
-                    color: Theme.of(
-                      context,
-                    ).colorScheme.shadow.withValues(alpha: 0.1),
-                    blurRadius: 20,
-                    offset: const Offset(0, 10),
-                  ),
-                ],
-              ),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(24),
-                child: Image.asset(
-                  personalInfo.profileImage,
-                  fit: BoxFit.cover,
-                  errorBuilder: (context, error, stackTrace) {
-                    return Container(
-                      decoration: BoxDecoration(
-                        color: Theme.of(context).colorScheme.primaryContainer,
-                        borderRadius: BorderRadius.circular(20),
-                      ),
-                      child: Icon(
-                        Icons.person,
-                        size: 100,
-                        color: Theme.of(context).colorScheme.onPrimaryContainer,
-                      ),
-                    );
-                  },
-                ),
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 20.0),
-            child: Text(
-              "Amarjit Mallick",
-              style: TextStyle(
-                fontSize: 42,
-                fontWeight: FontWeight.w800,
-                letterSpacing: 2,
-                color: Colors.white,
-              ),
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.symmetric(vertical: 20.0),
-            child: Text(
-              "A Software Engineer who has developed countless innovative solutions",
-              textAlign: TextAlign.center,
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 18,
-              ),
-            ),
-          ),
-          _buildSocialLinks(context),
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSocialLinks(
-    BuildContext context,
-  ) {
-    final socialLinks = [
-      {
-        'name': 'LinkedIn',
-        'icon': 'assets/icons/linkedin.png',
-        'url': 'https://www.linkedin.com/in/amarjit-mallick/',
-      },
-      {
-        'name': 'GitHub',
-        'icon': 'assets/icons/github.png',
-        'url': 'https://github.com/amarjitmallick',
-      },
-      {
-        'name': 'Twitter',
-        'icon': 'assets/icons/twitter.png',
-        'url': 'https://x.com/amarjitmallick_',
-      },
-    ];
-
-    return Container(
-      constraints: const BoxConstraints(maxWidth: 600),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            spacing: 16,
-            children: socialLinks.map((link) {
-              return IconButton(
-                onPressed: () {
-                  launchUrl(
-                    Uri.parse(link['url'] as String),
-                    mode: LaunchMode.externalApplication,
-                  );
-                },
-                icon: Image.asset(
-                  link['icon'] as String,
-                  width: 36,
-                  height: 36,
-                  fit: BoxFit.contain,
-                ),
-              );
-            }).toList(),
-          ),
-        ],
       ),
     );
   }
@@ -325,9 +221,13 @@ class PublicationsScreen extends StatelessWidget {
                       ),
                     ),
                     Text(
-                      DateFormat("MMM dd, yyyy").format(DateTime.parse(publication.year)),
+                      DateFormat(
+                        "MMM dd, yyyy",
+                      ).format(DateTime.parse(publication.year)),
                       style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                        color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.6),
+                        color: Theme.of(
+                          context,
+                        ).colorScheme.onSurface.withValues(alpha: 0.6),
                         fontStyle: FontStyle.italic,
                       ),
                     ),
@@ -338,7 +238,9 @@ class PublicationsScreen extends StatelessWidget {
                   publication.description,
                   maxLines: 1,
                   style: Theme.of(context).textTheme.bodyMedium?.copyWith(
-                    color: Theme.of(context).colorScheme.onSurface.withValues(alpha: 0.8),
+                    color: Theme.of(
+                      context,
+                    ).colorScheme.onSurface.withValues(alpha: 0.8),
                     fontWeight: FontWeight.w500,
                   ),
                 ),
@@ -349,7 +251,9 @@ class PublicationsScreen extends StatelessWidget {
                     icon: const Icon(Icons.read_more_rounded),
                     label: const Text('Read More'),
                     style: TextButton.styleFrom(
-                      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8)),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(8),
+                      ),
                       backgroundColor: Colors.white,
                       foregroundColor: Theme.of(context).colorScheme.onPrimary,
                     ),
