+++ lib/widgets/skills_section.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/material.dart';
+import 'package:portfolio/models/portfolio_models.dart';
+
+class SkillsSection extends StatelessWidget {
+  final List<Skill> skills;
+
+  const SkillsSection({
+    super.key,
+    required this.skills,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final groupedSkills = <String, List<Skill>>{};
+    for (final skill in skills) {
+      groupedSkills.putIfAbsent(skill.category, () => []).add(skill);
+    }
+
+    return Container(
+      decoration: BoxDecoration(
+        color: Theme.of(context).colorScheme.surfaceContainerHighest,
+        borderRadius: BorderRadius.circular(16),
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            'Technical Skills',
+            style: Theme.of(context).textTheme.titleLarge?.copyWith(
+              fontWeight: FontWeight.w600,
+              fontSize: 32,
+            ),
+          ),
+          const SizedBox(height: 24),
+          Table(
+            border: TableBorder.all(
+              color: Colors.white.withValues(alpha: 0.2),
+              borderRadius: BorderRadius.circular(8),
+            ),
+            children: [
+              ...groupedSkills.entries.map((entry) {
+                return TableRow(
+                  children: [
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 12,
+                        vertical: 12,
+                      ),
+                      child: Text(
+                        entry.key,
+                        style: Theme.of(context).textTheme.titleMedium
+                            ?.copyWith(fontWeight: FontWeight.w600),
+                      ),
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 12,
+                        vertical: 12,
+                      ),
+                      child: Row(
+                        spacing: 10,
+                        children: entry.value
+                            .map(
+                              (skill) => Text(
+                                skill.name,
+                                style: Theme.of(context).textTheme.titleMedium
+                                    ?.copyWith(fontWeight: FontWeight.w600),
+                              ),
+                            )
+                            .toList(),
+                      ),
+                    ),
+                  ],
+                );
+              }),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
