+++ lib/widgets/stat_card.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+import 'package:portfolio/models/portfolio_models.dart';
+
+class StatCard extends StatelessWidget {
+  final StatItem stat;
+  final bool isDesktop;
+
+  const StatCard({
+    super.key,
+    required this.stat,
+    this.isDesktop = true,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      decoration: BoxDecoration(
+        color: Theme.of(context).colorScheme.surfaceContainerHighest,
+        borderRadius: BorderRadius.circular(16),
+      ),
+      child: Column(
+        mainAxisAlignment: isDesktop
+            ? MainAxisAlignment.center
+            : MainAxisAlignment.start,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            stat.number,
+            style: Theme.of(context).textTheme.headlineMedium?.copyWith(
+              fontSize: 64,
+              fontWeight: FontWeight.bold,
+              color: Colors.white,
+            ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            stat.label,
+            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+              color: Theme.of(
+                context,
+              ).colorScheme.onSurface.withValues(alpha: 0.7),
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+}
