+++ pubspec.yaml
@@ -16,6 +16,7 @@ dependencies:
   http: ^1.5.0
   intl: ^0.20.2
   flutter_svg: ^2.2.3
+  font_awesome_flutter: ^10.12.0
 
 dev_dependencies:
   flutter_test:
